Bug fix - Issue with search results fixed
Fixed an issue caused by improper use of lodash union to show the search results along with the books in the existing state. Issue fixed by using map function along with lodash find. The existing state of the book is now visible in the search page as well. Minor modifications to code based on suggestions.